+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapter.kt
@@ -123,6 +123,9 @@ class ExoPlayerAdapter(
                     TAG,
                     "Collector was attached while media source was already playing, transitioning to playing state"
                 )
+
+                // We need to add at least one ms here to not run into videoStartupTime = 0
+                stateMachine.addStartupTime(1)
                 stateMachine.transitionState(PlayerStates.PLAYING, position)
             }
         }