+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - updates ExoPlayer to `2.18.X`
 
+### Fixed
+- fixed issue where late attaching of collector caused `videoStartupTime` = 0
+
 ## Development v1
 
 ## v2.9.0